moving stats functions
